+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.0.2 (work in progress)
 
 * ifdef preprocessor, uses [gpp](https://github.com/Kestrer/gpp)
+* added abstract console
  
 0.0.1